+++ src/main/java/fr/isika/projet3/service/PromoterService.java
@@ -7,7 +7,7 @@ import fr.isika.projet3.entities.Promoter;
 public interface PromoterService {
 
 	public List<Promoter> getAllPromoters();
-	public Promoter getPromoterById(int id);
+	public Promoter getPromoterById(Long id);
 	public boolean savePromoter(Promoter promoter);
-	public boolean deletePromoterById(int id);
+	public boolean deletePromoterById(Long id);
 }
